utils - apiFeatures.js: refactoring the API features
